+++ noodles-bgzf/CHANGELOG.md
@@ -1,6 +1,6 @@
 # Changelog
 
-## Unreleased
+## 0.10.0 - 2022-03-02
 
 ### Changed